python_api_service/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
ultralytics
torch
torchvision